Agrega validacion para no agregar a un usuario a un torneo inexistente

+++ backend/src/main/java/utn/frba/wordle/dto/MemberDto.java
@@ -10,5 +10,5 @@ import lombok.*;
 public class MemberDto {
 
     private String username;
-    private Integer tournamentId;
+    private Long tournamentId;
 }
